+++ SickBeard.py
@@ -36,6 +36,8 @@ import warnings
 
 warnings.filterwarnings('ignore', module=r'.*fuzzywuzzy.*')
 warnings.filterwarnings('ignore', module=r'.*Cheetah.*')
+warnings.filterwarnings('ignore', module=r'.*connectionpool.*', message='.*certificate verification.*')
+warnings.filterwarnings('ignore', module=r'.*ssl_.*', message='.*SSLContext object.*')
 
 if not (2, 7, 9) <= sys.version_info < (3, 0):
     print('Python %s.%s.%s detected.' % sys.version_info[:3])